requirements.txt
# Core dependencies
PyGObject>=3.42.0
Pillow>=9.0.0
pycairo>=1.20.0

# Google Gemini API
google-generativeai>=0.3.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0